got everything working except for editorSelectSyntaxHighlighting()
