+++ src/find.h
@@ -1,0 +1,7 @@
+#ifndef FIND_C_
+#define FIND_C_
+
+void editorFindCallback(char *query, int key);
+void editorFind();
+
+#endif
