+++ src/input.h
@@ -1,0 +1,8 @@
+#ifndef INPUT_C_
+#define INPUT_C_
+
+char *editorPrompt(char *prompt, void (*callback)(char *, int));
+void editorMoveCursor(int key);
+void editorProcessKeypress();
+
+#endif
