+++ src/output.h
@@ -1,0 +1,11 @@
+#ifndef OUTPUT_C_
+#define OUTPUT_C_
+
+
+void editorScroll();
+void editorDrawRows(struct abuf *ab);
+void editorDrawStatusBar(struct abuf *ab);
+void editorRefreshScreen();
+void editorSetStatusMessage(const char *fmt, ...);
+
+#endif
